+++ run_example.py
@@ -1,4 +1,3 @@
-import sys
 import time
 from pathlib import Path
 import os
@@ -8,7 +7,6 @@ from commonroad.common.file_reader import CommonRoadFileReader
 import commonroad_route_planner.fast_api.fast_api as fapi
 from commonroad_route_planner.reference_path import ReferencePath
 
-from commonroad_velocity_planner.configuration.optimization_config import JerkMinType
 # own code base
 from commonroad_velocity_planner.utils.visualization.visualize_velocity_planner import visualize_global_trajectory
 from commonroad_velocity_planner.utils.visualization.visualize_quantities import (
@@ -106,13 +104,9 @@ if __name__ == "__main__":
     scenarios = "/home/tmasc/projects/velocity_planner/commonroad-velocity-planner/scenarios"
     output_dir_path: str = "/home/tmasc/projects/velocity_planner/commonroad-velocity-planner/output"
     xml = scenarios + "/" +"ZAM_handling1-1_1_T-1.xml"
-    main(path_to_xml=xml, save_img=False, test=False, output_dir_path=output_dir_path)
-
-
-    sys.exit()
     for xml in sorted(os.listdir(scenarios)):
         _xml = scenarios + "/" + xml
-        main(path_to_xml=_xml, save_img=True, test=False, output_dir_path=output_dir_path)
+        main(path_to_xml=_xml, save_img=False, test=False, output_dir_path=output_dir_path)
 
     def test_cr_challenge():
         cr_challenge = "/home/tmasc/projects/route_planner/commonroad-reference_path-planner/tutorial/commonroad_challenge_2023"